Update patch set 2
Patch Set 2:

> I am just getting more confused :-)
 > 
 > Here is the latest https://github.com/alexbrainman/goissue14790/blob/master/20160325/transcript.txt

If we ignore the "With100usDelay" part for now, then I think situation becomes more clear. The osyield change helps, right? We expected that osyield is, in fact, a yield and not a 15ms delay (hopefully it was true at some point). And we sprinkled osyields in lots of places (e.g. mutex, GC and my change 20835 adds another one, well actually it replaces usleep(100) with osyield, but still). On your machine osyield is a 15ms sleep. This is very wrong and can leads to all kinds of slowdowns and unexpected latencies.

Please cook a patch that changes osyield to use SwitchToThread.

Now with cl/20835 and fixed osyield, we don't have any unexplained cases, right?
BenchmarkDefaultResolution is still slow (but not BenchmarkDefaultResolution-4!). That's because sysmon actually needs to retake P from blocked syscall and that is limited by system clock resolution. To mitigate this, we also need to land this change (reduces latency from 32ms to 2ms).

Patch-set: 2