crypto/tls: server accept ECDHE without curves or points
Prior to this change if a client sent ECDHE ciphers without the curve or
point extensions the go server would not accept those ciphers.

According to RFC 4492 section 4, that is not correct behavior:

"A client that proposes ECC cipher suites may choose not to include
these extensions.  In this case, the server is free to choose any one of
the elliptic curves or point formats listed in Section 5."

Change-Id: I22ba247e7bbb5cdf0ac8914f186f1ab31f57cb92