Update patch set 1
Patch Set 1:

> I'm not sure what you mean? Inflate uses io.ByteReader rather than io.Reader to do it's work.

And that's kinda the problem.

Part of io.Reader's contract is that you should return as much as possible from Read, without waiting.

ByteReader has no such contract. A bufio.Reader.ReadByte may be acting from memory, or may do a Read to get more data.

So implementing an io.Reader in terms of a ByteReader which in turn is implemented in terms of an io.Reader makes it hard to implement the io.Reader contract, since the visibility into "can that byte be fetched immediately" is lost at the ByteReader.

My question is: will this result in more Read calls (i.e. more network blocking or whatever) than before?

Really, it'd be nice if in the cases where we need to make our own byte reader (in newFlateReader or whatever), we'd also keep track of the bufio.Buffered() state, so we knew when it was safe to call newBits().

Maybe that's as simple as doing a .(*bufio.Reader).Buffered() > 0 before anything optimistically calling moreBits.

Patch-set: 1