Update patch set 1
Patch Set 1:

> trybot failure is real. investigating.

There's some code somewhere in iterating over strings that expects the element type of a string to be a byte.  I keep meaning to fix it, because it causes us to call typPtr(nil).

It ends up not causing problems in practice because we never dereference that pointer.

Patch-set: 1
Label: TryBot-Result=0