runtime: allocate black during GC
DO NOT SUBMIT. We're still thinking about this.

TODO: Update the big comment at the top of mgc.go and probably other
comments.

TODO: Benchmarks.

Currently we allocate white for most of concurrent GC based on the
classical argument that it produces less floating garbage than
allocating black. While this is true, allocating black has several
advantages that may significantly outweigh this downside.

1) It naturally bounds the total scan work to the live heap size at
the start of a GC cycle. Allocating white does not, and thus depends
entirely on assists to prevent the heap from growing faster than it
can be scanned.

2) It reduces the total amount of scan work per GC cycle by the size
of newly allocated objects that are linked into the heap graph, since
objects allocated black do not need to be scanned.

3) It reduces total write barrier work since more objects will already
be black when they are linked into the heap graph.

Change-Id: I2214a6a0e4f544699ea166073249a8efdf080dc0