Update patch set 2
Patch Set 2: Code-Review-2

If this change is submitted, it would be unprecedented: A feature added to Go corresponding to a feature being removed from C++. Trigraphs are scheduled to be removed in C++17. Of course, trigraphs have come and gone before, with IBM driving their retention, so the deletion might not land. The irony is thick.

As someone who learned Unix typing on an upper-case-only DECWRITER (a paper terminal), using # and @ for character erase and line kill, \## to erase a backslash, and \[ to type a brace, yet finding that freeing after programming on IBM OS/360 using a type 029 card keypunch (a big advance over the 026), I do not have fondness for the constraints of the past.

IBM has hurt me enough already.

Patch-set: 2
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2