cmd/compile: hoist induction variable
Remove the hidden induction variable when it's used
only for looping such as in:

for _, v := range slice {
}

The compiler generates a hidden iterator 0 -> len(slice),
which can be removed. Shaves about 1-2 bytes per loop.

There is no loop hoisting so the upperbound is computed
every iteration. There are some register spilling I don't
understand for example in (*Checker).packageObjects in
types/resolver.go.

TODO: Is it easier to just generate better code?

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I03162d878148ed6384cfa9e1bb481574189f2412