Update patch set 9
Patch Set 9:

Brad, that's what I rewrote.  It no longer barfs on the closure, but instead:

--- FAIL: TestDLLPreloadMitigation (0.20s)
	syscall_windows_test.go:844: Bad: insecure load of DLL by base name "nojack.dll" before sysdll registration: <nil>
--- FAIL: TestLoadLibraryEx (0.00s)
	syscall_windows_test.go:862: Expected LoadLibraryEx+flags to be available. (LoadLibraryEx=false; flags=false)

The old-style load preceding this one worked, but the secure load failed for some reason, and (in theory) I did rote changes to some closure code to only get rid of the memory allocation.

Patch-set: 9