Update patch set 1
Patch Set 1:

I could move the "Simple custom benchmarks that do not use the "go test command can also be run using the Benchmark function" line in testing.go all the way down after the description of the standard BenchmarkXxx way (i.e. move to to line 80). That would make it way less prominent.

As for the example, godoc will show it in the middle of the docs after the function signature (it's not one of those package-level examples rendered at the top) so I don't think it'll fool anyone into thinking that's the encouraged way to write benchmarks.

Patch-set: 1