Update patch set 1
Patch Set 1: Code-Review-2

Maybe for Go 1.6.2, but not Go 1.6.1 (which is a security-only release).

Just blocking this with -2 temporarily.

This would also need to be decided on Github or the mailing list about whether it goes into Go 1.6.x, not here on Gerrit.

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2