net: prevent potential race on access to internal API hooks in testing
This change introduces a naive giant lock that serializes execution of
tester goroutines using internal API hooks.

Perhaps the race might be a root cause of unpredictable dial i/o timeout
failures on buildbots and trybots. The failures appear to happen with
tip (or after the use of testing.Parallel.)

Change-Id: Ibfd6b55fe824ac96fe11f4c8fe2578be0338659b