Update patch set 1
Patch Set 1:

On amd64, on all operating systems, the top 16 bits are all either 0 on 1. So we can shift 15 bits, instead of 16 bits (essentially preserving the MSB), and use 15+3 bits instead of 16+3 bits for the ABA counter.

Then when we unpack, we sign extend, and it will work on all amd64 systems today and in the future, without having to special case Solaris.

Patch-set: 1
Reviewer: Gerrit User 5036 <5036@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: TryBot-Result=0