cmd/pprof: use symbol info from profiles
Pprof profiles may contain symbol info on the form:

if the profile was queried with ?debug=1.
Now pprof does not error out when binary is not specified,
so try to use that embed info when usual symbolization fails
(e.g. binary is not specified or any other error).
This allows to use self-contained profiles as:

$ go tool pprof --text profile.file

The motivation for this is tracer: change 21732 embeds symbol
info into traces. This is enough to visualize traces without binary.
But trace tool can aslo generate some pprof profiles and visualize
them using pprof tool. That won't work without binary. With with change
tracer will be able to emit symbol info into pprof profiles and
invoke pprof tool without binary.

Change-Id: I8381272027c7c9be4d83a01f9213479308c9a21d