Update patch set 4
Patch Set 4:

What are the implications of this change for compatibility between old trace programs and new binaries and vice versa? Hana made some comments that sound like they are not compatible. If the trace format is changing, there is a version string in the trace header that should be updated, so that the trace tool can continue to work with both old and new traces.

Patch-set: 4