Update patch set 1
Patch Set 1:

> (1 comment)

As a matter of practice, the open source and proprietary projects I'm generally involved with avoid putting issue numbers in source code since if you change bug tracking systems later, the issue# is now a dead pointer.

The contribution page also does not mention this guideline.

I'm happy to make the change, but if it is actually acceptable or expected to reference issue# in source code, it would be helpful if the contributions page mentioned it.

Do you still want this change?

Patch-set: 1
Label: Run-TryBot=0