Update patch set 6
Patch Set 6:

I have changed the code as discussed in the go-dev thread.

This means that only "level 1" is replaced. This is now the mode, where we do matching to the previous block, since it has very small speed penalty compared to not doing that.

I have updated both reference and new benchmarks to reflect the latest versions:

https://docs.google.com/spreadsheets/d/1VLxi-ac0BAtf735HyH3c1xRulbkYYUkFecKdLPH7NIQ/edit?usp=sharing

Patch-set: 6