Update patch set 1
Patch Set 1:

I'd prefer to not see code duplicated, so perhaps a newArrayAlloc() that doesn't do the division check, and alter newarray() to call that.
Then newarray's callers can be examined to see if it makes sense to convert any to the new function, e.g. they're already doing the check.

But I'm just an opinionated spectator, so wait for Ian, etc., to pass judgement before wasting your effort.  :-)

Patch-set: 1