Update patch set 1
Patch Set 1:

> (1 comment)

On my Intel Haswell benchmark pc there is no measureable difference between both versions even for the common case in the benchmark loop as those comparisons are cheap and usually predicted correctly (and if the if is taken then the program panics anyway). 

However, adding the != precondition will introduce some extra branch mispredictions when len=cap and len!=cap allocations are mixed. So i do not think this would be an improvement and i would like to leave it as is in this CL.

Patch-set: 1
Reviewer: Gerrit User 5846 <5846@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0