Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> (2 comments)
> 
> Any notable generated code changes now that this a hint rather than a requirement?

Not much size or performance wise.  The main visible difference is that before we did:
MOVQ AX, BX
ADDQ $5, BX
and AX contains the unmodified value.  Now we do
MOVQ AX, BX
ADDQ $5, AX
and BX contains the unmodified value.  To the extent that it matters (it might not, with the right register renaming hardware) the latter has a shorter critical path.

Patch-set: 1