Update patch set 1
Patch Set 1:

> Would be nice, if we had partial inlining, for checks like these in the beginning of a function/method. Could that be worth a proposal/issue?

It's not obvious to me how the compiler would decide how much to inline. And it might then need to keep around multiple copies of the inlined function, depending on how much got inlined. But certainly feel free to file an issue for discussion if you like; never hurts.

Patch-set: 1