Update patch set 1
Patch Set 1: Code-Review+2

Thanks.

To make you feel better about some of the lost time, switching to symbol generation was a precursor for moving a big chunk of the DWARF generation into the compiler, which means a lot of it will happen in parallel and not be stuck on the linker critical path.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2