Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Thanks. I agree with the change.

We should consider having the pprof package generate the profiles in profile.proto format, so that any adjustments are applied when the profile is generated, not when it is consumed.

Patch-set: 2
Reviewer: Gerrit User 10031 <10031@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1