Update patch set 5
Patch Set 5:

The code itself is in decent shape, but I still question the scope.

There are two mental models for certificate verification. Some libraries take the "word of god" approach: a certificate is "verified" and thereafter everything in it can be trusted by the application. On the other hand, some libraries (including Go) only verify certificates for a specific use (i.e. name).

Checking the directory name when a directory name isn't a verification target confuses the two models. However, the fundamental fault is already in the code: I wasn't clear enough about the model from the start and have struggled to align around a single view since given the API-compat requirements.

Having looked over the code again and done a quick search for how it's being used on GitHub, it looks like it would be safer to convert the existing DNS name constraint support into "word of god" mode since that's stricter.

Also, after having checked the Baseline document, it doesn't actually say that constrained intermediates must include limitations on the directory name, although it is in the example so general CA practice might be to include it.

Also, this change is intertwined with adding email addresses and IP addresses as verification targets. Maybe that makes sense given that these things are exposed in the Certificate structure, although it's making me regret exposing them at all in the first place.


So I'm going to poke at converting the existing constraints code into "word of god" mode and then circle back to this.

Patch-set: 5