crypto: align certificate verification around being target-specific.
There are two models for certificate verification. The original model
is that certificates are verified generically and, afterwards, treated
as the “word of god”. I.e. once verified, applications are free to
extract and trust any information from the certificate.

An alternative model is that verification is done for a specific target,
such as a domain name. In this model verification doesn't answer whether
a certificate is “verified” or not, rather it says whether the public
key should be trusted for a specific use or not.

The motivation for the latter model is that it admits more schemes. For
example, certificates authenticated by DNSSEC (should that ever take
off) cannot be generically validated because DNS is inheriently scoped
by domain name.

Change-Id: I5052279b98815a44cabde1fb4f2d27647b9e977b