Update patch set 5
Patch Set 5:

Paul: please see https://go-review.googlesource.com/#/c/21904. This hopefully makes the model of constraints more uniform in the code. I'm not wedded to the target-specific verification model, but it does seem to be more general. So I'm interested in any arguments for the "word of god" model before moving in that direction.

(Thankfully it's the case that trying to verify a certificate with DNSName constraints already fails unless a DNSName is specified for verification. So I just made that clearer in the code.)

Patch-set: 5
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0