Update patch set 1
Patch Set 1:

(7 comments)

While I like the changes in this code mostly it does not solve the limited support of certificate constraints.

Constraining certificates is intended to limit the “word of god” instead of applying it, making it more target specific. This change is still mostly ignoring constraints applied from the CA certificate(s) down to it's child's.

Name constraints are in principal are applied from a parent to leaf certificates. Where target specific name checks are applied to the leaf certificate only from the application level.

I don't see the relation with DNSSEC where similar constraints are applied 
naturally by the DNS hierarchy.

Patch-set: 1