Update patch set 1
Patch Set 1:

> While I like the changes in this code mostly it does not solve the
 > limited support of certificate constraints.
 > 
 > Constraining certificates is intended to limit the “word of god”
 > instead of applying it, making it more target specific. This change
 > is still mostly ignoring constraints applied from the CA
 > certificate(s) down to it's child's.

I think I've failed to explain this difference very well and perhaps you're actually arguing for a "word of god" model. (And maybe that's the right way to go.)

At the moment, constraints on the DNS names are applied to the VerifyOptions.DNSName value. Consider a leaf with SANs for "www.foo.com" and "www.bar.com" issued by an intermediate constrained to just ".foo.com".

If we tried to verify that for the name "www.foo.com" then, currently, that would succeed. In the target-specific model that chain is acceptable because we only care about whether a certificate is valid for the target in question.

A generic validation would have to reject that leaf because it contains a name that's not permitted by the constraints on the intermediate. A generic validation couldn't know what information the application will extract from the verified certificate so must check everything.

 > I don't see the relation with DNSSEC where similar constraints are
 > applied
 > naturally by the DNS hierarchy.

Consider a world where we trust DNSSEC signatures on DANE records. A certificate chain could then contain an X.509, leaf certificate followed by a bunch of DNS records and signatures to the DNS root.

In a target-specific world, we could verify such a certificate for www.foo.com if the DNS records and signatures chained through com, foo.com and www.foo.com.

However, a generic validation would have to be very restrictive. Just ensuring that the names in the cert matched the DNS path wouldn't be sufficient because some application could be assigning semantic meaning to some extension that we don't know about etc. It's fundamentally very difficult to implement a "word of god" model in this case.


So the question, which I'm not sure about the answer to, is whether we implement a "word of god" model or whether we aim at a target-specific model.

Patch-set: 1
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0