Update patch set 2
Patch Set 2:

I don't like this. The lists are long, but they are very easy to read. Usually there's some kind of obvious pattern to the TXXX cases too. Sometimes there's a pattern to Oxxx too. This CL changes long lists that are easy to read into blocks of text that are very hard to quickly scan by eye.

Also "single-column" lists, so to speak, are very easy to re-arrange. These are much harder to work with, if a pattern is supposed to be preserved.

To add on that it's harder to read a diff when adding or subtracting cases (I complained about this in the previous CL too).

Patch-set: 2
Reviewer: Gerrit User 5036 <5036@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0