Update patch set 1
Patch Set 1:

> At the moment, constraints on the DNS names are applied to the
 > VerifyOptions.DNSName value. Consider a leaf with SANs for
 > "www.foo.com" and "www.bar.com" issued by an intermediate
 > constrained to just ".foo.com".
 > 
 > If we tried to verify that for the name "www.foo.com" then,
 > currently, that would succeed. In the target-specific model that
 > chain is acceptable because we only care about whether a
 > certificate is valid for the target in question.

I don't have any real problem with this, it might be a performance consideration to not validate the whole certificate but only for what it's used. The problem is that the current code don't allow you to specify or validate a certificate for a different purpose.

 > A generic validation would have to reject that leaf because it
 > contains a name that's not permitted by the constraints on the
 > intermediate. A generic validation couldn't know what information
 > the application will extract from the verified certificate so must
 > check everything.

This would strengthen the security in my opinion.

Else you would trust a false issued certificate, the CA was never allowed to issue a certificate for www.bar.com in the first place. What does this say about the trustworthiness of the certificate in general?

 > Consider a world where we trust DNSSEC signatures on DANE records.
 > A certificate chain could then contain an X.509, leaf certificate
 > followed by a bunch of DNS records and signatures to the DNS root.
 > 
 > In a target-specific world, we could verify such a certificate for
 > www.foo.com if the DNS records and signatures chained through com,
 > foo.com and www.foo.com.
 > 
 > However, a generic validation would have to be very restrictive.
 > Just ensuring that the names in the cert matched the DNS path
 > wouldn't be sufficient because some application could be assigning
 > semantic meaning to some extension that we don't know about etc.
 > It's fundamentally very difficult to implement a "word of god"
 > model in this case.
 > 
 > 
 > So the question, which I'm not sure about the answer to, is whether
 > we implement a "word of god" model or whether we aim at a
 > target-specific model.

Personally I think it's a combination of the two.

1> We strictly validate though a generic model while we build/verify the certificate chain.
2> If the chain is established correctly we check if the certificate is allowed to be used for the specified target and/or purpose (this could also be done in the reverse order for performance reasons).

In the case of DNSSEC is the question what you want the certificate to proof. This would mostly depend on the application who queries for the certificate. Probably you do care about the subject of the certificate (validated organizational information) but don't care about the value of the subject alternative names as the domain control is proof is already done through DNSSEC. But even if you care about the SAN values they would still be populated according to the RFC and the constraints can still be validated.

If there are no constraints you would only validate the target. If there are conflicting constraints it was clearly not the intention for you the use the certificate in the way you are doing. In the case your certificate was issued by a trusted CA you might need to obtain a new certificate from them or when using in in house CA (untrusted) you need to modify your CA settings and re-issue.

Other non-standard certificate values belong in a extension. Specific restricted purposes can be defined in the extended key-usage.

Patch-set: 1