cmd/compile: intern Types to shrink LocalSlot size
This is semi-experimental; if this is done in roughly the
right way, if it is not too vile, then the next step would
be to do more pervasive interning of Type and to perhaps
take advantage of types being interned in other phases (like
CSE, for example).

In all of make.bash, the maximum number of distinct types
mentioned in a function is 75, so 32 bits for the type-intern
field is surely overkill.  Red-black trees were used for the
intern data structure because they were handy and because
Type already supports comparison.

Change-Id: I89d397a82e5a2f0aefaa731de156e843e2e44753