cmd/compile: switch to compact export format by default
The new format reduces the export data size by ~65% on average
for the std library packages.

The compiler transparently handles both packages using the old
and the new format.

How can you tell which one you have (if you care): Open a package
(.a) file in an editor. Textual export data starts with a $$ after
the header and is more or less legible; binary export data starts
with a $$B after the header and is mostly unreadable. A stand-alone
decoder (for debugging) is in the works.

Comparing the assembly code for every package of the std library
revealed no relevant changes in code (the assembly code is identical
for all but the syscall package, and there it's only two variables
that are swapped in use, resulting in identical semantics).

In case of a problem, please first try reverting back to the old
textual format to determine if the cause is the new export format:

For stand-alone compiler invocation:
- go tool compile -newexport=false <files>

For a single package:
- go build -gcflags="-newexport=false" <pkg>

For make/all.bash:
- (export GO_GCFLAGS="-newexport=false"; sh make.bash)

Fixes #13241.

Change-Id: I2588cb463be80af22446bf80c225e92ab79878b8