Update patch set 1
Patch Set 1:

This conflicts with a pending CL of mine (https://go-review.googlesource.com/c/22144/) which I'd like to get in first if possible.

Regarding testing: I'm not sure if I ever wrote up a proposal on the bug tracker, but I was planning on running a server(s) on GCE which just exists for network integration tests, acting like anything we want it to act like for the purpose of Go's tests. Want to help with that during the freeze period?

Patch-set: 1