Update patch set 2
Patch Set 2:

> Agree. This is a breaking change that should not be submitted. The
 > difference from behavior in C is unfortunate but unavoidable now.

ok. will abandon this but it might be good if we clarify the behavior in the fmt documentation instead. The 0x does not count towards the width for padding with zeros #06x but it counts for padding with spaces #6x which might be surprising.

Patch-set: 2
Reviewer: Gerrit User 5846 <5846@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0